Update Player still cant move

+++ Assets/Script/Player_Script/PlayerController.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public class PlayerController : MonoBehaviour
 {
-    [SerializeField] public float speed;
+    [SerializeField] public int speed;
 }
